requirements.txt
# Python 3.11 only
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0
pydantic==2.7.1

# Orchestration
langgraph==0.2.28
langchain-core==0.2.38

# Database clients
neo4j==5.20.0
chromadb==0.5.3
aiosqlite==0.20.0

# ML / Detection
scikit-learn==1.4.2
shap==0.45.1
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
chronos-forecasting==1.3.0
transformers==4.41.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
aiofiles==23.2.1
structlog==24.1.0
cryptography==42.0.7

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
